+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Removed
 
+## [0.0.9] - 2024-02-21
+
+### Fixed
+
+- macos build is now fixed for the sys crate.
+
 ## [0.0.8] - 2024-01-31
 
 ### Added
@@ -95,7 +101,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation of zones, frames, plots & messages
 
-[unreleased]: https://github.com/den-mentiei/tracy-gizmos/compare/v0.0.8...HEAD
+[unreleased]: https://github.com/den-mentiei/tracy-gizmos/compare/v0.0.9...HEAD
+[0.0.9]: https://github.com/den-mentiei/tracy-gizmos/compare/v0.0.8..v0.0.9
 [0.0.8]: https://github.com/den-mentiei/tracy-gizmos/compare/v0.0.7..v0.0.8
 [0.0.7]: https://github.com/den-mentiei/tracy-gizmos/compare/v0.0.6..v0.0.7
 [0.0.6]: https://github.com/den-mentiei/tracy-gizmos/compare/v0.0.5..v0.0.6